xen: actually setup hypercalls.
This was somehow dropped during the iterations of the original Xen patches.
It's actually harmless at the moment since there are no users of hypercalls but
patches are being written to support Xen PV block devices which need this.

It's not clear exactly how early this needs to be but I think it needs to be at
least before init_hw() (since that would detect disk devices).

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>