vgabios: Minor fixes to struct vbe_info definition.
Although the VBE3 spec says there is 189 bytes of pad at the end of
the vbe_info struct, the spec also says the total struct size should
be 256 bytes.  The VBE2 spec uses sufficient pad to get to 256 bytes.
So, assume the spec was off by one and make the pad 190 bytes.

Also, use u32 for capabilities and 'struct segoff_s' for
win_func_ptr.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>