vgabios: Don't call int10 during mode switch.
Don't call int10 to load the fonts - instead call the font loading
functions directly.

Theoretically, this could change behavior if an external program has
captured int10 and redirects the font loading calls.  However, there
does not seem to be any indication that an external app could expect
that to work.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>