Use inline functions instead of macros for GET/SET_SEG segment manipulation.
It appears that gcc does a better job of optimization when the
SET_SEG() segment assignment assembler code is contained in an inline
function.  With the code in a function gcc appears to be able to
optimize out many redundant segment register loads.  Removing some of
these unnecessarily loads makes the code both smaller and faster.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>