Remove disk GET/SET_INT13DPT and GET/SET_INT13EXT macros.
The code generation is better if explicit 'struct int13ext_s' and
'struct int13dpt_s' pointers are used instead.  The code is a little
easier to understand as well.

This patch also forces disk_1348 to not be inlined.  If it gets
inlined into its caller it can increase the stack usage for all disk
calls.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>