update bios date
Linux ignores some information from acpi in case the bios is old
as acpi support used to be buggy in the early days.  This affects
ressources for example (unless forced with pci=use_crs).  So lets
go for something more recent.  With this patch applied the
ressources for \\SB.PCI0 will show up in /proc/{iomem,ioports},
tagged as "PCI Bus 0000:00".

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>