Remove SMBIOS generation on coreboot hack.
Early versions of coreboot did not create an SMBIOS table (which Linux
needs in order to use the ACPI tables), so SeaBIOS had a hack to
generate an SMBIOS table.  However, recent versions of coreboot can
generate an SMBIOS table, so there is no longer a reason to support
this hack.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>