acpi: LNKS is not needed
LNKS is a bit strange in that it reuses the same PIIX register as LNKA,
but has a different interrupt.  This means that the _CRS it returns will
not be one of the possible resources from _PRS.  This shows up in the
Linux boot logs as

   ACPI: PCI Interrupt Link [LNKS] (IRQs 9) *0

Instead of that, we can simply use a hardwired interrupt index.

Cc: Gleb Natapov <gnatapov@redhat.com>
Cc: Laszlo Ersek <lersek@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>