POST: Move cpu caching and dma setup to platform_hardware_setup().
The CPU cache is enabled and DMA is disabled on all real-world POST
entry situations, so no need to do this in the low-level "pre-init"
phase.  Instead, move it to the platform hardware setup stage.

Also, move the setting of the reboot flags (which control reset-vector
entry point handling) to ivt_init().

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>