Only perform SMP setup on QEMU.
The SMP setup initializes MTRRs on extra CPUs and populates the
CountCPUs variable.  On coreboot the CPUs should already be
initialized and the CountCPUs variable isn't used anywhere.

There have also been reports of the SIPI does not work on some real
machines.  So, it's best not to invoke one if it isn't needed.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>