Run ahci code entirely in 32bit mode.
The ahci driver needs to jump into 32bit mode in order to access
portions of the ahci controllers PCI config space.  Instead of jumping
into 32bit mode just to toggle the ahci registers, jump into 32bit
mode for all of the driver interactions.  This shrinks the size of the
overall code and can lead to further cleanups.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>