vgahooks: add SM720 VGA BIOS hooks for WIN Enterprises MB-60470
Add vgahooks to support the SMI SM720 VGA BIOS used on the
WIN Enterprises MB-60470.  The response from smi_157f14() is necessary
for the CRT output to turn on.  The response from smi_157f02() is used
to configure which connector(s) output is routed to.

As I lack hardware to test LCD panel output, I've selected CRT-only
output on the MB-60470 as this prevents the CRT output from being scaled
to the resolution of the LCD panel.

Signed-off-by: Jonathan A. Kollasch <jakllsch@kollasch.net>