Clear the total block transfer count on error.
The total number of blocks transferred is part of the return status of
the BIOS API.  This is normally the total number of blocks requested.
However, in an error it is possible for this count to be less.  Add a
global check to ensure that on an error event the block count is
updated - if it hasn't been updated, assume that no blocks have been
transferred successfully (the common case).

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>