build: Don't trash the .config file when src/Kconfig changes.
Revision 5325e91e introduced a change to force the rebuild of the
.config file when the src/Kconfig (or vgasrc/Kconfig) file changes.
However, this was causing a complete rebuild of the user's .config
file which would throw away the user's current settings.

On a rebuild, use the kbuild olddefconfig rule instead of defconfig to
attempt to keep the user's current settings.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>